+++ middleware.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/jexia/semaphore/pkg/broker"
-	"github.com/jexia/semaphore/pkg/broker/config"
 	"github.com/jexia/semaphore/pkg/flow"
 	"github.com/jexia/semaphore/pkg/functions"
 	"github.com/jexia/semaphore/pkg/references"
@@ -12,17 +11,17 @@ import (
 )
 
 // WithMiddleware initialises the given middleware and defines all options
-func WithMiddleware(middleware config.Middleware) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func WithMiddleware(middleware Middleware) Option {
+	return func(ctx *broker.Context, options *Options) {
 		options.Middleware = append(options.Middleware, middleware)
 	}
 }
 
-// BeforeConstructor the passed function gets called before new specifications are constructed
-func BeforeConstructor(wrapper config.BeforeConstructorHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+// WithBeforeConstructor the passed function gets called before new specifications are constructed
+func WithBeforeConstructor(wrapper BeforeConstructorHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.BeforeConstructor == nil {
-			options.BeforeConstructor = wrapper(func(*broker.Context, functions.Collection, config.Options) error { return nil })
+			options.BeforeConstructor = wrapper(func(*broker.Context, functions.Collection, Options) error { return nil })
 			return
 		}
 
@@ -30,9 +29,9 @@ func BeforeConstructor(wrapper config.BeforeConstructorHandler) config.Option {
 	}
 }
 
-// AfterConstructor the passed function gets called once all options have been applied
-func AfterConstructor(wrapper config.AfterConstructorHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+// WithAfterConstructor the passed function gets called once all options have been applied
+func WithAfterConstructor(wrapper AfterConstructorHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.AfterConstructor == nil {
 			options.AfterConstructor = wrapper(func(*broker.Context, specs.FlowListInterface, specs.EndpointList, specs.ServiceList, specs.Schemas) error {
 				return nil
@@ -45,8 +44,8 @@ func AfterConstructor(wrapper config.AfterConstructorHandler) config.Option {
 }
 
 // BeforeManagerDo the passed function gets called before a request gets handled by a flow manager
-func BeforeManagerDo(wrapper flow.BeforeManagerHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func BeforeManagerDo(wrapper flow.BeforeManagerHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.BeforeManagerDo == nil {
 			options.BeforeManagerDo = wrapper(func(ctx context.Context, manager *flow.Manager, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -60,8 +59,8 @@ func BeforeManagerDo(wrapper flow.BeforeManagerHandler) config.Option {
 }
 
 // BeforeManagerRollback the passed function gets called before a rollback request gets handled by a flow manager
-func BeforeManagerRollback(wrapper flow.BeforeManagerHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func BeforeManagerRollback(wrapper flow.BeforeManagerHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.BeforeManagerRollback == nil {
 			options.BeforeManagerRollback = wrapper(func(ctx context.Context, manager *flow.Manager, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -75,8 +74,8 @@ func BeforeManagerRollback(wrapper flow.BeforeManagerHandler) config.Option {
 }
 
 // AfterManagerDo the passed function gets after a flow has been handled by the flow manager
-func AfterManagerDo(wrapper flow.AfterManagerHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func AfterManagerDo(wrapper flow.AfterManagerHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.AfterManagerDo == nil {
 			options.AfterManagerDo = wrapper(func(ctx context.Context, manager *flow.Manager, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -90,8 +89,8 @@ func AfterManagerDo(wrapper flow.AfterManagerHandler) config.Option {
 }
 
 // AfterManagerRollback the passed function gets after a flow rollback has been handled by the flow manager
-func AfterManagerRollback(wrapper flow.AfterManagerHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func AfterManagerRollback(wrapper flow.AfterManagerHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.AfterManagerRollback == nil {
 			options.AfterManagerRollback = wrapper(func(ctx context.Context, manager *flow.Manager, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -105,8 +104,8 @@ func AfterManagerRollback(wrapper flow.AfterManagerHandler) config.Option {
 }
 
 // BeforeNodeDo the passed function gets called before a node is executed
-func BeforeNodeDo(wrapper flow.BeforeNodeHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func BeforeNodeDo(wrapper flow.BeforeNodeHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.BeforeNodeDo == nil {
 			options.BeforeNodeDo = wrapper(func(ctx context.Context, node *flow.Node, tracker flow.Tracker, processes *flow.Processes, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -120,8 +119,8 @@ func BeforeNodeDo(wrapper flow.BeforeNodeHandler) config.Option {
 }
 
 // BeforeNodeRollback the passed function gets called before a node rollback is executed
-func BeforeNodeRollback(wrapper flow.BeforeNodeHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func BeforeNodeRollback(wrapper flow.BeforeNodeHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.BeforeNodeRollback == nil {
 			options.BeforeNodeRollback = wrapper(func(ctx context.Context, node *flow.Node, tracker flow.Tracker, processes *flow.Processes, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -135,8 +134,8 @@ func BeforeNodeRollback(wrapper flow.BeforeNodeHandler) config.Option {
 }
 
 // AfterNodeDo the passed function gets called after a node is executed
-func AfterNodeDo(wrapper flow.AfterNodeHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func AfterNodeDo(wrapper flow.AfterNodeHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.AfterNodeDo == nil {
 			options.AfterNodeDo = wrapper(func(ctx context.Context, node *flow.Node, tracker flow.Tracker, processes *flow.Processes, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -149,8 +148,8 @@ func AfterNodeDo(wrapper flow.AfterNodeHandler) config.Option {
 }
 
 // AfterNodeRollback the passed function gets called after a node rollback is executed
-func AfterNodeRollback(wrapper flow.AfterNodeHandler) config.Option {
-	return func(ctx *broker.Context, options *config.Options) {
+func AfterNodeRollback(wrapper flow.AfterNodeHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
 		if options.AfterNodeRollback == nil {
 			options.AfterNodeRollback = wrapper(func(ctx context.Context, node *flow.Node, tracker flow.Tracker, processes *flow.Processes, store references.Store) (context.Context, error) {
 				return ctx, nil
@@ -159,5 +158,19 @@ func AfterNodeRollback(wrapper flow.AfterNodeHandler) config.Option {
 		}
 
 		options.AfterNodeRollback = wrapper(options.AfterNodeRollback)
+	}
+}
+
+// AfterFlowConstructor the passed function gets called before a flow manager gets constructed
+func AfterFlowConstructor(wrapper AfterFlowConstructionHandler) Option {
+	return func(ctx *broker.Context, options *Options) {
+		if options.AfterNodeRollback == nil {
+			options.AfterFlowConstruction = wrapper(func(*broker.Context, specs.FlowInterface, *flow.Manager) error {
+				return nil
+			})
+			return
+		}
+
+		options.AfterFlowConstruction = wrapper(options.AfterFlowConstruction)
 	}
 }
