+++ pkg/broker/providers/providers.go
@@ -1,8 +1,8 @@
 package providers
 
 import (
+	"github.com/jexia/semaphore"
 	"github.com/jexia/semaphore/pkg/broker"
-	"github.com/jexia/semaphore/pkg/broker/config"
 	"github.com/jexia/semaphore/pkg/checks"
 	"github.com/jexia/semaphore/pkg/compare"
 	"github.com/jexia/semaphore/pkg/dependencies"
@@ -26,7 +26,7 @@ type Collection struct {
 // The specifications are received from the providers. The property types are
 // defined and functions are prepared. Once done is a specs collection returned
 // that could be used to update the listeners.
-func Resolve(ctx *broker.Context, mem functions.Collection, options config.Options) (Collection, error) {
+func Resolve(ctx *broker.Context, mem functions.Collection, options semaphore.Options) (Collection, error) {
 	if options.BeforeConstructor != nil {
 		err := options.BeforeConstructor(ctx, mem, options)
 		if err != nil {
